Fix ScalaTest swallowing tests
