+++ tests/src/test/scala/neotypes/fs2/Fs2Suite.scala
